+++ BriansUsbQuizBoxApi/QuizBoxApi.cs
@@ -19,6 +19,8 @@ namespace BriansUsbQuizBoxApi
 
         private bool _idleMode = false;
 
+        private byte? _protocolVersion = null;
+
         private readonly WinnerByteSM _winnerByteSM;
         private readonly StatusByteSM _statusByteSM;
         private readonly GameStatusByteSM _gameStatusByteSM;
@@ -33,6 +35,9 @@ namespace BriansUsbQuizBoxApi
         private int? _threadId;
 
         #region Events
+
+        /// <inheritdoc/>
+        public event EventHandler<ConnectionCompleteEventArgs>? ConnectionComplete;
 
         /// <inheritdoc/>
         public event EventHandler<BuzzInEventArgs>? BuzzIn;
@@ -85,8 +90,8 @@ namespace BriansUsbQuizBoxApi
             _gameStatusByteSM = new GameStatusByteSM(
                 () => GameStarted?.Invoke(this, null),
                 () => GameLightOn?.Invoke(this, null),
-                (p, r1, r2, r3, r4, g1, g2, g3, g4) => GameDone?.Invoke(this, new GameDoneEventArgs(p, r1, r2, r3, r4, g1, g2, g3, g4)),
-                (p, r1, r2, r3, r4, g1, g2, g3, g4) => BuzzInStats?.Invoke(this, new BuzzInStatsEventArgs(p, r1, r2, r3, r4, g1, g2, g3, g4))
+                (p1, p2, p3, p4, p5, p6, p7, p8, r1, r2, r3, r4, g1, g2, g3, g4) => GameDone?.Invoke(this, new GameDoneEventArgs(p1, p2, p3, p4, p5, p6, p7, p8, r1, r2, r3, r4, g1, g2, g3, g4)),
+                (p1, p2, p3, p4, p5, p6, p7, p8, r1, r2, r3, r4, g1, g2, g3, g4) => BuzzInStats?.Invoke(this, new BuzzInStatsEventArgs(p1, p2, p3, p4, p5, p6, p7, p8, r1, r2, r3, r4, g1, g2, g3, g4))
             );
         }
 
@@ -96,9 +101,16 @@ namespace BriansUsbQuizBoxApi
             get { return _api.IsConnected; }
         }
 
+        /// <inheritdoc/>
         public QuizBoxTypeEnum? ConnectedQuizBoxType
         {
             get { return _api.ConnectedQuizBoxType; }
+        }
+
+        /// <inheritdoc/>
+        public byte? ProtocolVersion
+        {
+            get { return _protocolVersion; }
         }
 
         /// <inheritdoc/>
@@ -422,6 +434,14 @@ namespace BriansUsbQuizBoxApi
 
         private void ProcessRead(BoxStatusReport status, bool checkCommandWrite)
         {
+            bool bConnectionComplete = false;
+
+            if (_protocolVersion.HasValue == false)
+            {
+                _protocolVersion = status.ProtocolVersion;
+                bConnectionComplete = true;
+            }
+
             _winnerByteSM.Process(status.Status, status.Winner);
 
             _statusByteSM.Process(status.Status);
@@ -460,6 +480,11 @@ namespace BriansUsbQuizBoxApi
             else
             {
                 _idleMode = false;
+            }
+
+            if (bConnectionComplete)
+            {
+                ConnectionComplete?.Invoke(this, new ConnectionCompleteEventArgs(_protocolVersion.Value));
             }
         }